testy cd

+++ program/include/test.hpp
@@ -4,7 +4,7 @@
 
 #include "prediction.hpp"
 
-void test(string expected_series_name, int expected_next_number, unique_ptr<Prediction> actual);
+void test(string expected_series_name, float expected_next_number, unique_ptr<Prediction> actual);
 void run_tests();
 
 #endif
